refactor: update sbi-testing，简化测试
