+++ test-kernel/Cargo.toml
@@ -8,7 +8,7 @@ publish = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-sbi-testing = { git = "https://github.com/rustsbi/sbi-testing", rev = "fb260eb" }
+sbi-testing = { git = "https://github.com/rustsbi/sbi-testing", rev = "b64abbe" }
 riscv = "0.8"
 spin = "0.9"
 r0 = "1"
